==> 0_exploration.Rmd <==
---
title: "Olympic EDA and preprocessing"
output: html_notebook
---

KEY Findings:
- rimosso "Art competition" dagli sport
- shifta l'anno delle olimpiadi invernali di due, cosi' da avere il medesimo anno per estate/inverno
- 75% missing prima del 1960 in altezza e peso --> ~20% dopo il 1960. Che fare??

```{r}
library("plotly")
library("tidyverse")
library("data.table")
library("gridExtra")
library("knitr")
library("maps")
```

```{r}
# Load athletes_events data 
data <- read_csv("input/athlete_events.csv",
                 col_types = cols(
                   ID = col_character(),
                   Name = col_character(),
                   Sex = col_factor(levels = c("M","F")),
                   Age =  col_integer(),
                   Height = col_double(),
                   Weight = col_double(),
                   Team = col_character(),
                   NOC = col_character(),
                   Games = col_character(),
                   Year = col_integer(),
                   Season = col_factor(levels = c("Summer","Winter")),
                   City = col_character(),
                   Sport = col_character(),
                   Event = col_character(),
                   Medal = col_factor(levels = c("Gold","Silver","Bronze"))
                 )
)

noc <- read_csv("input/noc_regions.csv",
                col_types = cols(
                  NOC = col_character(),
                  region = col_character()
                ))

data_regions <- data %>% 
  left_join(noc, by = "NOC") %>%
  filter(!is.na(region))
```


Come e' cambiato il numero di nazioni, atleti, sport ed eventi nel corso del tempo?
```{r}
# L.A 1932: The Games were held during the worldwide Great Depression and 
# many nations and athletes were unable to pay for the trip to Los Angeles
# Melbourne 1952: alcune nazioni, come la Svizzera, i Paesi Bassi, la Spagna 
# boicottarono i Giochi in segno di protesta contro l'Unione Sovietica
# Montreal 1976: Attendance dipped because 25 nations, mostly African, 
# boycotted the Games in reponse to apartheid policies in South Africa.
# Mosca 1980: Olimpiade del boicottaggio americano per l'invasione sovietica 
# dell'Afghanistan, esempio seguito da altre 64 nazioni.
# count number of athletes, nations, & events, excluding the Art Competitions
counts <- data %>% filter(Sport != "Art Competitions") %>%
  group_by(Year, Season) %>%
  summarize(
    Athletes = length(unique(ID)),
    Nations = length(unique(NOC)),
    Events = length(unique(Event)),
    Sports = length(unique(Sport))
  )

# plot
p1 <- ggplot(counts, aes(x=Year, y=Athletes, group=Season, color=Season)) +
  geom_point(size=2) +
  geom_line() +
  scale_color_manual(values=c("darkorange","darkblue")) +
  xlab("") +  
  annotate("text", x=c(1932,1956,1976,1980),
           y=c(2000,2750,6800,4700),
           label=c("L.A. 1932","Melbourne 1956","Montreal 1976","Moscow 1980"),
           size=3) +
  annotate("text",x=c(1916,1942),y=c(10000,10000),
           label=c("WWI","WWII"), size=4, color="red") +
  geom_segment(mapping=aes(x=1914,y=8000,xend=1918,yend=8000),color="red", size=2) +
  geom_segment(mapping=aes(x=1939,y=8000,xend=1945,yend=8000),color="red", size=2)
p2 <- ggplot(counts, aes(x=Year, y=Nations, group=Season, color=Season)) +
  geom_point(size=2) +
  geom_line() +
  scale_color_manual(values=c("darkorange","darkblue")) +
  xlab("") +  
  annotate("text", x=c(1932,1976,1980),
           y=c(60,105,70),
           label=c("L.A. 1932","Montreal 1976","Moscow 1980"),
           size=3)
p3 <- ggplot(counts, aes(x=Year, y=Events, group=Season, color=Season)) +
  geom_point(size=2) +
  geom_line() +
  scale_color_manual(values=c("darkorange","darkblue"))
p4 <- ggplot(counts, aes(x = Year, y = Sports, group = Season, color = Season)) +
  geom_point(size = 2) +
  geom_line() +
  scale_color_manual(values = c("darkorange", "darkblue"))
grid.arrange(p1, p2, p3, p4, ncol=2)
```

- Arte ai giochi olimpici
L'arte ha fatto parte del programma dei Giochi olimpici dall'edizione del 1912 svoltasi a Stoccolma all'edizione 1948 tenutasi a Londra. Le gare di arte furono accantonate nel 1954 perché gli artisti venivano ormai considerati professionisti, a differenza degli atleti, che erano allora obbligati ad essere dilettanti. Dal 1956 venne istituito il programma culturale olimpico.
Numero di eventi, nazioni e artisti nel corso del tempo.
```{r}
# Subset to Art Competitions and variables of interest
art <- data %>% 
  filter(Sport == "Art Competitions") %>%
  select(Name, Sex, Age, Team, NOC, Year, City, Event, Medal)

# Count Events, Nations, and Artists each year
counts_art <- art %>% filter(Team != "Unknown") %>%
  group_by(Year) %>%
  summarize(
    Events = length(unique(Event)),
    Nations = length(unique(Team)),
    Artists = length(unique(Name))
  )

# Create plots
p5 <- ggplot(counts_art, aes(x=Year, y=Events)) +
  geom_point(size=2) +
  geom_line() + xlab("")
p6 <- ggplot(counts_art, aes(x=Year, y=Nations)) +
  geom_point(size=2) +
  geom_line() + xlab("")
p7 <- ggplot(counts_art, aes(x=Year, y=Artists)) +
  geom_point(size=2) +
  geom_line()
grid.arrange(p5, p6, p7, ncol=1)
```
Quali Paesi hanno vinto di piu' nelle competizioni artistiche?
```{r}
# count number of medals awarded to each Team
medal_counts_art <- art %>% filter(!is.na(Medal))%>%
  group_by(Team, Medal) %>%
  summarize(Count=length(Medal)) 

# order Team by total medal count
levs_art <- medal_counts_art %>%
  group_by(Team) %>%
  summarize(Total=sum(Count)) %>%
  arrange(Total) %>%
  select(Team)
medal_counts_art$Team <- factor(medal_counts_art$Team, levels=levs_art$Team)

# plot
ggplot(medal_counts_art, aes(x=Team, y=Count, fill=Medal)) +
  geom_col() +
  coord_flip() +
  scale_fill_manual(values=c("gold1","gray70","gold4")) +
  ggtitle("Historical medal counts from Art Competitions") +
  theme(plot.title = element_text(hjust = 0.5))
```
It is remarkable that Germany won the most medals in the Art Competitions between 1912 and 1948, considering that Germany was not invited to participate in 3 of the 7 Olympics during this period (they were banned from the 1920, 1924, and 1948 Olympics due to post-war politics). However, Germany made up for these absences with an especially strong showing at the 1936 Berlin Olympics, a.k.a., the Nazi Olympics, in which they won 40% of the medals in the Art Competitions and 60% of all the Art Competition medals in the country’s history.

Dominazione nazista nelle olimpiadi di Berlino del 1936
```{r}
# count number of medals awarded to each Team at Nazi Olympics
medal_counts_art_nazis <- art %>% filter(Year==1936, !is.na(Medal))%>%
  group_by(Team, Medal) %>%
  summarize(Count=length(Medal)) 

# order Team by total medal count
levs_art_nazis <- medal_counts_art_nazis %>%
  group_by(Team) %>%
  summarize(Total=sum(Count)) %>%
  arrange(Total) %>%
  select(Team)
medal_counts_art_nazis$Team <- factor(medal_counts_art_nazis$Team, levels=levs_art_nazis$Team)

# plot
ggplot(medal_counts_art_nazis, aes(x=Team, y=Count, fill=Medal)) +
  geom_col() +
  coord_flip() +
  scale_fill_manual(values=c("gold1","gray70","gold4")) +
  ggtitle("Nazi domination of Art Competitions at the 1936 Olympics") +
  theme(plot.title = element_text(hjust = 0.5))
```

Escludiamo l'arte dagli sport presenti nel dataset.
```{r}
data <- data %>% filter(Sport != "Art Competitions")
```

A partire dal 1992 le olimpiadi invernali si tengono due anni prima di quelle estive, ricodifichiamo quindi gli anni
per queste olimpiadi invernali cosi' da avere un unico anno per olimpiade.
```{r}
# Recode year of Winter Games after 1992 to match the next Summer Games
# Thus, "Year" now applies to the Olympiad in which each Olympics occurred 
original <- c(1994,1998,2002,2006,2010,2014)
new <- c(1996,2000,2004,2008,2012,2016)
for (i in 1:length(original)) {
  data$Year <- gsub(original[i], new[i], data$Year)
}
data$Year <- as.integer(data$Year)
```

Save dataframe
```{r}
df <- data %>% 
  left_join(noc, by = "NOC") %>%
  filter(!is.na(region))
save(df, file = "input/data.rdata")
```

Partecipazione maschile/femminile nel corso del tempo.
```{r}
# Table counting number of athletes by Year and Sex
counts_sex <- data %>% group_by(Year, Sex) %>%
  summarize(Athletes = length(unique(ID)))
counts_sex$Year <- as.integer(counts_sex$Year)

# Plot number of male/female athletes vs time
ggplot(counts_sex, aes(x=Year, y=Athletes, group=Sex, color=Sex)) +
  geom_point(size=2) +
  geom_line()  +
  scale_color_manual(values=c("darkblue","red")) +
  labs(title = "Number of male and female Olympians over time") +
  theme(plot.title = element_text(hjust = 0.5))
```

Proporzione tra il numero di donne e di uomini (linea tratteggiata indica meta' donne e meta' uomini)
```{r}
# Count M/F/Total per country per Olympics 
# Keep only country-years with at least 50 athletes
counts_NOC <- data %>% filter(Year %in% c(1936,1956,1976,1996,2016)) %>%
  group_by(Year, NOC, Sex) %>%
  summarize(Count = length(unique(ID))) %>%
  spread(Sex, Count) %>%
  mutate(Total = sum(M,F,na.rm=T)) %>%
  filter(Total > 49)
names(counts_NOC)[3:4] <- c("Male","Female")
counts_NOC$Male[is.na(counts_NOC$Male)] <- 0
counts_NOC$Female[is.na(counts_NOC$Female)] <- 0
counts_NOC$Year <- as.factor(counts_NOC$Year)

# Plot female vs. male athletes by NOC / Year
ggplot(counts_NOC, aes(x=Male, y=Female, group=Year, color=Year)) +
  geom_point(alpha=0.6) +
  geom_abline(intercept=0, slope=1, linetype="dashed") +
  geom_smooth(method="lm", se=FALSE) +
  labs(title = "Female vs. Male Olympians from participating NOCs") +
  theme(plot.title = element_text(hjust = 0.5)) +
  guides(color=guide_legend(reverse=TRUE))
```

Rappresentazione geografica della partecipazione alle olimpiadi di
- Amsterdam 1928
- Munich 1972
- Rio 2016
```{r}
# Subset to Games of interest and count athletes from each country
amsterdam <- data_regions %>% 
  filter(Games == "1928 Summer") %>%
  group_by(region) %>%
  summarize(Amsterdam = length(unique(ID)))
munich <- data_regions %>% 
  filter(Games == "1972 Summer") %>%
  group_by(region) %>%
  summarize(Munich = length(unique(ID)))
rio <- data_regions %>% 
  filter(Games == "2016 Summer") %>%
  group_by(region) %>%
  summarize(Rio = length(unique(ID)))

# Create data for mapping
world <- map_data("world")
mapdat <- tibble(region = unique(world$region))
mapdat <- mapdat %>% 
  left_join(amsterdam, by="region") %>%
  left_join(munich, by="region") %>%
  left_join(rio, by="region")
mapdat$Amsterdam[is.na(mapdat$Amsterdam)] <- 0
mapdat$Munich[is.na(mapdat$Munich)] <- 0
mapdat$Rio[is.na(mapdat$Rio)] <- 0
world <- left_join(world, mapdat, by="region")
```
Amsterdam 1928
```{r}
# Plot: Amsterdam 1928
ggplot(world, aes(x = long, y = lat, group = group)) +
  geom_polygon(aes(fill = Amsterdam)) +
  labs(title = "Amsterdam 1928",
       x = NULL, y=NULL) +
  theme(axis.ticks = element_blank(),
        axis.text = element_blank(),
        panel.background = element_rect(fill = "navy"),
        plot.title = element_text(hjust = 0.5)) +
  guides(fill=guide_colourbar(title="Athletes")) +
  scale_fill_gradient(low="white",high="red")

```
Munich 1972
```{r}
# Plot: Munich 1972
ggplot(world, aes(x = long, y = lat, group = group)) +
  geom_polygon(aes(fill = Munich)) +
  labs(title = "Munich 1972",
       x = NULL, y = NULL) +
  theme(axis.ticks = element_blank(),
        axis.text = element_blank(),
        panel.background = element_rect(fill = "navy"),
        plot.title = element_text(hjust = 0.5)) +
  guides(fill=guide_colourbar(title="Athletes")) +
  scale_fill_gradient2(low = "white", high = "red")
```
Rio 2016
```{r}
# Plot:  Rio 2016
ggplot(world, aes(x = long, y = lat, group = group)) +
  geom_polygon(aes(fill = Rio)) +
  labs(title = "Rio 2016",
       x = NULL, y = NULL) +
  theme(axis.ticks = element_blank(),
        axis.text = element_blank(),
        panel.background = element_rect(fill = "navy"),
        plot.title = element_text(hjust = 0.5)) +
  guides(fill=guide_colourbar(title="Athletes")) +
  scale_fill_gradient2(low="white",high = "red")
```


Completezza dei dati su altezza e peso
```{r}
# Check data availability
par(mfrow=c(2,2))
data %>% group_by(Year, Sex) %>%
  summarize(Present = length(unique(ID[which(!is.na(Height) & !is.na(Weight))])),
            Total = length(unique(ID))) %>%
  mutate(Proportion = Present/Total) %>%
  ggplot(aes(x=Year, y=Proportion, group=Sex, color=Sex)) +
  geom_point() +
  geom_line() +
  scale_color_manual(values=c("darkblue","red"))  +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(title="Height/Weight data completeness from each Olympiad")

data %>% group_by(Year, Sex) %>%
  summarize(Present = length(unique(ID[which(!is.na(Height) & !is.na(Weight) & !is.na(Medal))])),
            Total = length(unique(ID[which(!is.na(Medal))]))) %>%
  mutate(Proportion = Present/Total) %>%
  ggplot(aes(x=Year, y=Proportion, group=Sex, color=Sex)) +
  geom_point() +
  geom_line() +
  scale_color_manual(values=c("darkblue","red"))  +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(title="Winners' height/weight data completeness from each Olympiad")

data %>% group_by(Year, Sex) %>%
  summarize(Present = length(unique(ID[which(!is.na(Height))])),
            Total = length(unique(ID))) %>%
  mutate(Proportion = Present/Total) %>%
  ggplot(aes(x=Year, y=Proportion, group=Sex, color=Sex)) +
  geom_point() +
  geom_line() +
  scale_color_manual(values=c("darkblue","red"))  +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(title="Height data completeness from each Olympiad")

data %>% group_by(Year, Sex) %>%
  summarize(Present = length(unique(ID[which(!is.na(Weight))])),
            Total = length(unique(ID))) %>%
  mutate(Proportion = Present/Total) %>%
  ggplot(aes(x=Year, y=Proportion, group=Sex, color=Sex)) +
  geom_point() +
  geom_line() +
  scale_color_manual(values=c("darkblue","red"))  +
  theme(plot.title = element_text(hjust = 0.5)) +
  labs(title="Weight data completeness from each Olympiad")
par(mfrow=c(1,1))
```

```{r}
# Remove missing Height/Weight data and limit to years from 1960 onward
data <- data %>% filter(!is.na(Height), !is.na(Weight), Year > 1959) 
```


Height and weight over time
```{r}
par(mfrow=c(2,1))
data %>% ggplot(aes(x=as.factor(Year), y=Height, fill=Sex)) +
  geom_boxplot(alpha=0.75) +
  xlab("Olympiad Year") + ylab("Height (cm)") +
  scale_fill_manual(values=c("blue","red"))
data %>% ggplot(aes(x=as.factor(Year), y=Weight, fill=Sex)) +
  geom_boxplot(alpha=0.75) +
  xlab("Olympiad Year") + ylab("Weight (kg)") +
  scale_fill_manual(values=c("blue","red"))
par(mfrow=c(1,1))
```

Out of 489 events that appeared at least once in the Olympics since 1960, only 136 were included in every Olympiad. These include events for both men and women in alpine skiing, athletics, canoeing, diving, equestrian, fencing, figure skating, gymnastics, speed skating, and swimming, as well as events for men only in basketball, biathlon, boxing, crosscountry skiing, cycling, football, field hockey, ice hockey, pentathlon, rowing, ski jumping, water polo, weightlifting, and wrestling.
```{r}
# Identify events present in all 15 Games
events <- data[data$Year==1960,"Event"] %>% unique %>% .$Event # 177 in 1960
years <- data$Year %>% unique %>% sort %>% tail(-1)
for (i in 1:length(years)) {
  nxt <- data[data$Year==years[i],"Event"] %>% unique %>% .$Event
  events <- intersect(events, nxt)
}

# Subset data to only these events
data <- data %>% filter(Event %in% events)

# Get list of sports matching events
sports_events <- data %>% select(Sport, Event) %>% unique
```
This is a lot of events to consider, but we can reduce the list a bit. First, we can eliminate events based on weight classes (wrestling, weightlifting, and boxing), since the size of athletes in these events are restricted and the changes over time primarily reflect shifting definitions of the weight classes. Second, we can eliminate events that include both men and women, including all the equestrian events and pairs figure skating. This leaves 108 events to consider.

To charaterize historical trends in size for different events, I fit separate linear regressions for Height ~ Year and Weight ~ Year for athletes in each event, and saved the estimated regression slopes. By plotting the estimated regression slopes for height against the estimated regression slopes for weight across different events, we can identify events in which the size of athletes have changed the most. Importantly, the quadrant of the plot in which the point falls indicates the type of size change for each event:

Upper left quadrant: athletes have gotten shorter and heavier
Upper right quadrant: athletes have gotten taller and heavier
Lower right quadrant: athletes have gotten taller and lighter
Lower left quadrant: athletes have gotten shorter and lighter
Here is an interactive plot for men’s events.

Change in height vs change in weight over time across men’s sports
```{r}
# Eliminate wrestling, weightlifting, and boxing
sports_events <- sports_events %>% 
  filter(!Sport %in% c("Wrestling","Weightlifting","Boxing","Equestrianism")) %>%
  filter(!Event %in% c("Figure Skating Mixed Pairs")) %>%
  arrange(Sport)

# Add column for men/women/mixed
sports_events$Sex <- ifelse(grepl("Women",sports_events$Event),"Women","Men")

# Loop through events and fit regressions
s.height <- s.weight <- c()
for (i in 1:nrow(sports_events)) {
  temp <- data %>% filter(Event == sports_events$Event[i])
  lm.height <- lm(Height ~ Year, data=temp)
  lm.weight <- lm(Weight ~ Year, data=temp)
  s.height[i] <- lm.height$coefficients["Year"]
  s.weight[i] <- lm.weight$coefficients["Year"]
}
slopes <- tibble(Sport = sports_events$Sport, 
                 Event = sports_events$Event,
                 Sex = sports_events$Sex,
                 Height = s.height,
                 Weight = s.weight)

# Multiple slopes by 56 since 56 years passed between 1960 to 2016
slopes$Height <- round(slopes$Height*56,1)
slopes$Weight <- round(slopes$Weight*56,1)

# Plot regression slopes of weight ~ height for men
g2.m <- ggplot(slopes[slopes$Sex=="Men",], aes(x=Height, y=Weight, color=Sport, label=Event)) +
  geom_point(alpha=0.75) +
  geom_vline(xintercept = 0) +
  geom_hline(yintercept = 0) +
  labs(title="Temporal trends in men's size in different events",
       x="Height (cm)",
       y="Weight (kg)")  +
  theme(plot.title = element_text(hjust = 0.5),
        legend.position="none")
ggplotly(g2.m)

# Plot regression slopes of weight ~ height for women
g2.f <- ggplot(slopes[slopes$Sex=="Women",], aes(x=Height, y=Weight, color=Sport, label=Event)) +
  geom_point(alpha=0.75) +
  geom_vline(xintercept = 0) +
  geom_hline(yintercept = 0) +
  labs(title="Temporal trends in women's size in different events",
       x="Height (cm)",
       y="Weight (kg)")  +
  theme(plot.title = element_text(hjust = 0.5),
        legend.position="none")
ggplotly(g2.f)
```

Possono essere assegnate due o piu' medaglie per lo stesso evento (cioe' in caso di parita')?
```{r}
df %>% filter(!is.na(Medal)) %>% group_by(Games, Event, Medal) %>% summarise(n = n()) %>% filter(n>1)
# Si', in questo caso un oro, un argento e due bronzi nella corsa 100m.
filter(df, Games == "1896 Summer" & Event == "Athletics Men's 100 metres")

```

Esistono poi eventi per cui vengono assegnate piu' di tre medaglie, per esempio in Tennis Doubles.
Quali sono gli eventi/sport che hanno assegnato piu' medaglie ?
Quali sono gli eventi che mediamente assegnano piu' medaglie ?
```{r}
df %>% filter(!is.na(Medal)) %>% group_by(Games, Event) %>% summarise(n = n()) %>% arrange(desc(n))
medals <- df %>% filter(!is.na(Medal)) %>% group_by(Games, Event) %>% 
  summarise(sum_medal = n()) %>% arrange(desc(sum_medal))
(medals_per_event <- medals %>% group_by(Event) %>% summarise(avg_medal = mean(sum_medal)) %>%
  arrange(desc(avg_medal)))

# Possiamo dedurre che per gli eventi per cui ci sono quattro o cinque medaglie ci sia un pareggio
(pareggi <- filter(medals, sum_medal %in% c(4, 5)))
# si ma che ce frega?!
```
